+++ docs/app/[[...slug]]/(home)/components/apps/studio.png
