+++ docs/app/[[...slug]]/(home)/components/hero/index.tsx
