+++ docs/app/[[...slug]]/(home)/components/installer.tsx
