+++ docs/app/[[...slug]]/(home)/components/video.tsx
