+++ docs/app/[[...slug]]/(home)/index.tsx
@@ -1,10 +1,10 @@
-import { Apps } from '../../components/apps';
-import { CallToAction } from '../../components/cta';
-import { Features } from '../../components/features';
-import { Hero } from '../../components/hero';
-import { OpenSource } from '../../components/open-source';
-import { Review } from '../../components/review';
-import { Social } from '../../components/social';
+import { Apps } from './components/apps';
+import { CallToAction } from './components/cta';
+import { Features } from './components/features';
+import { Hero } from './components/hero';
+import { OpenSource } from './components/open-source';
+import { Review } from './components/review';
+import { Social } from './components/social';
 
 const Home = () => (
   <main className="divide-y px-0 pt-[var(--fd-nav-height)]">
